Rename base game class to base-game
