+++ src/games/anagram/anagram.js
@@ -4,7 +4,7 @@ const inquirer = require('inquirer');
 const chalk = require('chalk');
 const clui = require('clui');
 const readline = require('readline');
-const Game = require('../abstract-game');
+const Game = require('../base-game');
 const anagramWordList = require('./anagram-word-list');
 const anagramConstants = require('./anagram-constants');
 require('events').EventEmitter.defaultMaxListeners = 100;
