+++ src/games/hangman/hangman.js
@@ -1,7 +1,7 @@
 const clear = require('clear');
 const figlet = require('figlet');
 const chalk = require('chalk');
-const Game = require('../abstract-game');
+const Game = require('../base-game');
 const hangmanStages = require('./hangman-stages');
 const HangmanPrompts = require('./hangman-prompts');
 const wordList = require('../word-list');
